move to util instead
