+++ packages/opds-web-client/src/hooks/useDownloadButton.ts
@@ -2,12 +2,7 @@ import { MediaLink, FulfillmentLink, MediaType } from "./../interfaces";
 import { useActions } from "../components/context/ActionsContext";
 import download from "../components/download";
 import { generateFilename, typeMap } from "../utils/file";
-
-export function fixMimeType(mimeType: MediaType): MediaType {
-  return mimeType === "vnd.adobe/adept+xml"
-    ? "application/vnd.adobe.adept+xml"
-    : mimeType;
-}
+import { fixMimeType } from "../utils/book";
 
 function isIndirect(
   link: MediaLink | FulfillmentLink
@@ -21,20 +16,12 @@ function isIndirect(
 export const STREAMING_MEDIA_LINK_TYPE: MediaType =
   "text/html;profile=http://librarysimplified.org/terms/profiles/streaming-media";
 
-const isMac = navigator.platform.indexOf("Mac") > -1;
-function isPlatformCompatible(link: MediaLink | FulfillmentLink) {
-  if (isMac && fixMimeType(link.type) === "application/vnd.adobe.adept+xml")
-    return false;
-  return true;
-}
-
 type DownloadDetails = {
   fulfill: () => Promise<void>;
   isIndirect: boolean;
   downloadLabel: string;
   mimeType: MediaType;
   fileExtension: string;
-  isPlatformCompatible: boolean;
   isStreaming: boolean;
 };
 /**
@@ -100,7 +87,6 @@ export default function useDownloadButton(
     isStreaming,
     downloadLabel,
     mimeType: mimeTypeValue,
-    fileExtension,
-    isPlatformCompatible: isPlatformCompatible(link)
+    fileExtension
   };
 }
